+++ src/plugin.js
@@ -40,12 +40,13 @@ const plugin = {
       ],
       since: "0.6.0"
     },
-    sortAttributesByKey: {
+    xmlSortAttributesByKey: {
       type: "boolean",
       category: "XML",
       default: false,
-      description: "Orders XML attributes by key alphabetically while prioritizing xmlns attributes.",
-    },
+      description:
+        "Orders XML attributes by key alphabetically while prioritizing xmlns attributes."
+    }
   },
   defaultOptions: {
     printWidth: 80,