+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
-- the `sortAttributesByKey: true | false` option has been added. See the README.
+- the `xmlSortAttributesByKey: true | false` option has been added. See the README.
 
 ## [3.0.0-alpha.0] - 2023-06-02